requirements.txt
openai
python-dotenv
openai-whisper
git+https://github.com/openai/whisper.git